Remove circular dependency from <mwc-radio> by merging selection-controller into base.
Also fix some typings.

PiperOrigin-RevId: 259059580
